Added .localhost name constraint to CA certificate
The name constraint extension that has been added to the CA certificate
should mean that the certificate is only trusted for domains that end in
`.localhost`. This has the benefit that the certificate cannot be used
to intercept connections to other domains, and the drawback that name
constraints are not well supported and mkcert can no longer issue certificates
to non-`.localhost` domains.