Add DNS whitelist option to CA generation
The -permit option can be used when the CA is generated to only allow
certain domains to have certificates issued by the local CA. This allows
for a development CA to have the minimum possible responsibility,
provided the dev cert stack checks the domain restrictions.

This uses the NameConstraints X509 extension, so it is not implemented
by all OSes/browsers, which is reflected in the help message for the
option.